Revert "Improve UI and move folder"
This reverts commit c6c79106bb5ab694b70b862feb88562b33f9460f.
