+++ src/main/java/seedu/address/logic/commands/controller/ConfirmationWindowController.java
@@ -1,4 +1,4 @@
-package seedu.address.ui.controller;
+package seedu.address.logic.commands.controller;
 
 import seedu.address.ui.ConfirmationWindow;
