+++ src/test/java/seedu/address/logic/commands/ClearCommandTest.java
@@ -5,7 +5,7 @@ import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.ui.controller.ConfirmationBypassController;
+import seedu.address.logic.commands.controller.ConfirmationBypassController;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
