+++ src/test/java/seedu/address/logic/commands/DeleteCommandTest.java
@@ -13,7 +13,7 @@ import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.ui.controller.ConfirmationBypassController;
+import seedu.address.logic.commands.controller.ConfirmationBypassController;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
